### 2018-02-20 #589 @IbrahimTanyalcin
Hi Arfon,

Thank you very much for the inquiry. I will directly give examples from a project where I am the frontend developer:

http://deogen2.mutaframe.com/

This work has been published in Nucleic Acids Research (https://academic.oup.com/nar/article/45/W1/W201/3819233) and in a newer version we are making several calls to other databases, while having to instantiate many different modules sequentially. Taskq helps with controlling the load order without bloating the page with polyfills while executing functions in a non-render blocking way (because it uses ES6 Promises combined with requestAnimationFrame.)

Combining Promises with requesAnimationFrame (rAF) in fact stems from an old concept of throttling browser events which I talk about this concept in one my latest work:

http://ieeexplore.ieee.org/document/8291800/

The article above is mainly about rendering performance and different visualization modules (https://github.com/IbrahimTanyalcin/LEXICON) which can benefit from taskq.

As far as I know, combining Promises with requestAnimationFrame has not been tried in javascript module proposals. It enables 4 things:

- Non-render blocking execution --> performance
- Fall back mechanism to rAF for older browsers (this is important because whenever I demonstrate a software to hospitals, researchers etc, there are always people stuck with ie10 etc. because of regulations etc.) --> compatibility
- Requires no polyfill/transpiling while taking advantage of latest EcmaScript implementation (Promises) - Control on browser execution of scripts --> debugging etc.

Let me know if you need further clarification,

Regards,
