### 2018-02-22 543698a
added example for Medium.com
