### 2018-03-11 #589 @arfon
@IbrahimTanyalcin - thanks for the additional context. I think this is pretty close to an edge-case for us (as to whether this is research software) but given your role (and it's usage) on the http://deogen2.mutaframe.com project I'm happy to proceed with this submission.

@raganwald - :wave: hi! Would you be willing to review this submission for JOSS? You can get an idea of what this means here: http://joss.theoj.org/about#reviewer_guidelines
