### 2018-03-26 #589 @arfon
@HadrienG, @xirdneh, @serine - would you be willing to review this submission for JOSS? Ideally I'd like to find two reviewers for this submission.
