### 2018-04-16 #589 @arfon
@fil, @pgadige - would you be willing to review this submission for JOSS?
