+++ example/exampleGooglePSI/nonRenderBlocking_conventional/script1.js
@@ -1,0 +1,3 @@
+!function(){
+	texts.push("script1");
+}()
