+++ example/exampleGooglePSI/renderBlocking/script2.js
@@ -1,0 +1,3 @@
+!function(){
+	texts.push("script2");
+}()
