+++ example/exampleGooglePSI/nonRenderBlocking_conventional/script0.js
@@ -1,0 +1,4 @@
+!function(){
+	texts = [];
+	texts.push("script0");
+}()
