### 2018-04-19 #589 @serine
sorry, I wouldn't be able to review this one. From some brief investigation I think it'll be pretty difficult for me to asses this one. For example the demo link (http://deogen2.mutaframe.com/) has a lot going on there in terms of javascript

![screenshot from 2018-04-19 11-44-22](https://user-images.githubusercontent.com/6700099/38966750-1b62c1bc-43c7-11e8-9ee0-51c9e29da6ce.png)

Also I didn't look particular hard but I didn't see this line in the demo, which is how you get the library from my brief read

	<script type="text/javascript" src="../taskq.js" charset="UTF-8"></script>

The README looks comprehensive, but I'm still not sure how can I make an assessment of how installable and runable this tool is.

All the best
