### 2018-04-19 #589 @IbrahimTanyalcin
Hi Serine,

I have 2 versions of the server, for the new features, I test them for android, iOS, Safari , Firefox and Chrome First in test subdomain.

For the test server, please visit test-tube.mutaframe.com

You can there see taskq in action. Many thanks!

IBT,
