### 2018-04-30 #589 @xirdneh
Hi, @arfon are you still looking for reviewers?
The topic seems interesting. I've worker with javascript async capabilities before.
Also, seems like the repository has a test suite and a `package.json` for easy installation.

I'm not entirely sure how the demo project uses it (http://deogen2.mutaframe.com/) since most of the functionality is distributed in multiple files. I'm having some problems finding out the entry point.

There are a fair amount of examples in the repository so I should be able to install and test the library locally.
