### 2018-06-15 #589 @arfon
@IbrahimTanyalcin - I'm afraid to say I don't think we're going to be able to review this submission in JOSS. I've tried multiple times to find reviewers (outside of this thread) to no avail. :-(
