### 2018-06-15 #780 @whedon
[REVIEW]: Async Modules Supporting ES5 & ES6 with Control Flow
**Submitting author:** @IbrahimTanyalcin (<a href="http://orcid.org/0000-0003-0327-5096">Ibrahim  Tanyalcin</a>)
**Repository:** <a href="https://github.com/IbrahimTanyalcin/taskq" target ="_blank">https://github.com/IbrahimTanyalcin/taskq</a>
**Version:** v2.1.4
