### 2018-06-15 #589 @arfon
@xirdneh - please head over to #780 for the actual review
