### 2018-06-29 #780 @arfon
> Is it because of the white spaces/- in the paper.md ?

Looks like this is because of the `&` in the title. I can't seem to figure out how to get Pandoc to escape it properly. A quick fix would be to change `&` to `and`.
