### 2018-06-29 #780 @IbrahimTanyalcin
I guess &amp; would work. Let's give it a try, I'll push paper.md again
