+++ taskq.js
@@ -24,7 +24,7 @@
     }
 }(this,function(window,document){
 	//version
-	var version = "2.2.3";
+	var version = "2.3.0";
 	//a salt for getters/setters
 	var salt = Math.random();
 	//current script
@@ -325,7 +325,7 @@
 						bKEC = b0._taskqKEC === undefined ? b0._taskqKEC = !!~keywordEnd.indexOf(bId) : b0._taskqKEC,
 						aC = aKSC && (aKEC || aL.some(function(d,i){return d === bId || (tasksMap[d]._taskqWaitFor && ~tasksMap[d]._taskqWaitFor.indexOf(bId));})),
 						bC = bKSC && (bKEC || bL.some(function(d,i){return d === aId || (tasksMap[d]._taskqWaitFor && ~tasksMap[d]._taskqWaitFor.indexOf(aId));}));
-					return aC*base+aL.length - bC*base - bL.length;
+					return aC*base + aL.length - bC*base - bL.length;
 				}).map(function(d,i){
 					return d[0];
 				});
@@ -363,7 +363,7 @@
 						bKEC = b._taskqKEC === undefined ? b._taskqKEC = !!~keywordEnd.indexOf(bId) : b._taskqKEC,
 						aC = aKSC && (aKEC || aL.some(function(d,i){return d === bId || (tasksMap[d]._taskqWaitFor && ~tasksMap[d]._taskqWaitFor.indexOf(bId));})),
 						bC = bKSC && (bKEC || bL.some(function(d,i){return d === aId || (tasksMap[d]._taskqWaitFor && ~tasksMap[d]._taskqWaitFor.indexOf(aId));}));
-					if (aC*base+aL.length - bC*base - bL.length > 0) {
+					if (!this.areCircular(a,b,tasksMap) && (aC*base + this.dependencyIsSubset(a,b,tasksMap)*aL.length - bC*base - bL.length > 0)) {
 						tasks[i] = b;
 						tasks[j] = a;
 						var windowStart = Math.max(0,j-8);
@@ -373,11 +373,12 @@
 						--i;
 						break inner;
 					}
-				}
-			}
+				};
+			};
 			
 			report ? console.log("Semi stable sorting done in: "+ steps + " steps, ~" +(Date.now()-start)+"ms") : void(0);
 			//console.log(tasks.slice());
+			this.clearDependencyLedger(tasks);
 			return tasks;
 		};
 		/*requestAnimationFrame (rAF) is used instead of Promise wrapper in older browsers (ie9+)*/
@@ -489,6 +490,116 @@
 				container.appendChild(script);
 			}
 			return thenable;
+		};
+		prt.dependencyIsSubset = function(a,b,tasksMap,checked){
+			a._taskqDependencyIsSubset = a._taskqDependencyIsSubset || [];
+			a._taskqDependencyIsNotSubset = a._taskqDependencyIsNotSubset || [];
+			if(~a._taskqDependencyIsSubset.indexOf(b)) {
+				return true;
+			} else if (~a._taskqDependencyIsNotSubset.indexOf(b)) {
+				return false;
+			}
+			var checked = checked || [],
+				aL = a._taskqWaitFor || prt.emptyArr,
+				bL = b._taskqWaitFor || prt.emptyArr;
+			if(!aL.length){
+				a._taskqDependencyIsNotSubset.push(b);
+				return false;
+			} else if (!bL.length) {
+				a._taskqDependencyIsNotSubset.push(b);
+				checked.push(b);
+				return false;
+			} else if (~checked.indexOf(b)) {
+				return;
+			} else if (!aL.map(function(d,i){return bL.indexOf(d)}).some(function(d,i){return !~d})) {
+				a._taskqDependencyIsSubset.push(b);
+				checked.push(b);
+				return true;
+			} else {
[…]
