+++ example/declerativeDependenciesExample/B.js
@@ -4,7 +4,6 @@
 		div.textContent = "B\n\u21e9";
 		taskq.export({},"exportB");
 	}
-	window.ass = document.currentScript;
 	X._taskqId = document.currentScript.dataset.taskqid;
 	X._taskqWaitFor = document.currentScript.dataset.taskqwaitfor.split(",").filter(d=>d);
 	taskq.push(X);
